+++ changelog.md
@@ -1,5 +1,12 @@
 # The Modding Tree changelog:
 
+## v2.3.6 - 
+- Performance improvements.
+- Fixed tooltips overlapping with the top display.
+- Added support for bulk challenge completions.
+- "Best" is updated automatically.
+- Fixed keeping Decimal values on reset.
+
 ## v2.3.5 - 12/21/20
 - Added resetTime, which tracks the time since a layer prestiged or was reset.
 - A layer node will be highlighted red if one of its subtabs is highlighted red.
@@ -8,7 +15,7 @@
 - Other minor fixes.
 
 ## v2.3.4 - 12/16/20
-- Added an node image feature.
+- Added a node image feature.
 - Resource display now always shows the amount of the currency the layer's gain is based on.
 - Added spacing between tree nodes.
 - Another attempt to fix tooltip flickering.